MAINT 1 rename enable.qt4 -> enable.qt 2 add import path changinvg
