+++ enable/qt/base_window.py
