+++ enable/qt/constants.py
