+++ enable/qt/image.py
