+++ enable/qt/qpainter.py
