+++ enable/qt/scrollbar.py
